Create demo register data for Schools, Addresses, PostTowns and Postcodes using edubase data.
The processing retrieves basic school data for each row in edubase.csv. Then we attempt to match but postcode
and lat long to uprn. If a match is found the school record is writter to schools.tsv, the address is written to
addresses.tsv, postcode to postcodes.tsv and postTown to postowns.tsv.

The tsv  files can then be loaded to the respective demo registers.